Make isEnabled() readable and make getKeyMappings() check for local
Some implementaions of ConfigTable allow the children to change
the columns, so the patch that added the useEffect() to monitor children
was needed. However this update caused too many updates to localStorage,
reading and setting it so frequently it was useless. This now makes more
thorough checks of local storage as it adds columns to component states.